+++ CHANGELOG.md
@@ -1,9 +1,16 @@
 # Change Log
 
+## 0.0.3 - 23 November 2021
+- Preview release
+  - Update Snippets
+  - Bugfix onEnterRules
+  - Add property, visual-bigquery.defaultTimezone
+
+
 ## 0.0.2 - 22 November 2021
-- Update Snippets
-- Add onEnterRules
-- 
+- Preview release
+  - Update Snippets
+  - Add onEnterRules
 
 
 ## 0.0.1 - 22 July 2021